add backend checks for user and session id

+++ app/controllers/beneficiaries_controller.rb
@@ -7,25 +7,37 @@ class BeneficiariesController < ApplicationController
         render json: beneficiary, status: :ok
     rescue ActiveRecord::RecordNotFound
         render json: ['Beneficiary not found'], status: :not_found
-        
-        
     end
 
     def create
-        beneficiary = Beneficiary.create!(beneficiary_params)
-        render json: beneficiary, status: :created
+        matching_household = Household.find(params[:household_id])
+        if session[:user_id] == matching_household.id
+            beneficiary = Beneficiary.create!(beneficiary_params)
+            render json: beneficiary, status: :created
+        else
+            render json: ['You must be responsible for this household to add a beneficiary']
+        end
     end
 
     def update 
         beneficiary = Beneficiary.find(params[:id])
-        beneficiary.update!(beneficiary_params)
-        render json: beneficiary, status: :accepted
+            if session[:user_id] == beneficiary.household.user.id
+            beneficiary.update!(beneficiary_params)
+            render json: beneficiary, status: :accepted
+        else
+            render json: ['You must be responsible for this household to edit beneficiary information']
+        end
     end
 
     def destroy
-        beneficiary = Beneficiary.find(params[:id])
-        beneficiary.destroy
-        head :no_content
+        matching_household = Household.find(params[:id])
+        if session[:user_id] == matching_household.id
+            beneficiary = Beneficiary.find(params[:id])
+            beneficiary.destroy
+            head :no_content
+        else
+            render json: ['You must be responsible for this household to delete this beneficiary']
+        end
     end
 
     private
